requirements.txt
# Core Framework
fastapi==0.118.3
uvicorn==0.37.0
starlette==0.48.0
pydantic==2.12.0
pydantic_core==2.41.1
python-multipart==0.0.20

# Database & ORM (PostgreSQL support)
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Redis (for job queue)
redis==5.0.1
hiredis==2.3.2

# HTTP & API
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
urllib3==2.5.0
certifi==2025.10.5
idna==3.10
charset-normalizer==3.4.3

# OpenAI API
openai==2.2.0
jiter==0.11.0

# PDF Processing
pdfplumber==0.11.7
pdfminer.six==20250506
pypdfium2==4.30.0
pytesseract==0.3.13
Pillow==10.4.0

# Data Processing
pandas==2.3.3
numpy==2.3.3
openpyxl==3.1.5
et_xmlfile==2.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Templates & Web
Jinja2==3.1.6
MarkupSafe==3.0.3

# Configuration
python-dotenv==1.1.1
click==8.3.0

# Utilities
tqdm==4.67.1
six==1.17.0
packaging==25.0
colorama==0.4.6
distro==1.9.0
cffi==2.0.0
pycparser==2.23
cryptography==46.0.2
annotated-types==0.7.0
anyio==4.11.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
